Provide resources/guidance on latitude/longitude
In the past, developers spent too much time trying to figure out how
lat/long worked, to the point that they didn't get around to writing
much Mongoose code. This commit adds link to resources on lat/long
as well as hemispheres of the earth, and specifies that lat/long should
be numbers as opposed to strings.

Close #78.